Move miscellaneous capability code to a dedicated file
