+++ caps/capabilities_test.go
@@ -23,8 +23,6 @@ import (
 	"testing"
 
 	. "gopkg.in/check.v1"
-
-	"github.com/ubuntu-core/snappy/testutil"
 )
 
 func Test(t *testing.T) {
@@ -34,25 +32,6 @@ func Test(t *testing.T) {
 type CapabilitySuite struct{}
 
 var _ = Suite(&CapabilitySuite{})
-
-func (s *CapabilitySuite) TestValidateName(c *C) {
-	c.Assert(ValidateName("name with space"), ErrorMatches,
-		`"name with space" is not a valid snap name`)
-	c.Assert(ValidateName("name-with-trailing-dash-"), ErrorMatches,
-		`"name-with-trailing-dash-" is not a valid snap name`)
-	c.Assert(ValidateName("name-with-3-dashes"), IsNil)
-	c.Assert(ValidateName("name"), IsNil)
-}
-
-func (s *CapabilitySuite) TestLoadBuiltInTypes(c *C) {
-	repo := NewRepository()
-	err := LoadBuiltInTypes(repo)
-	c.Assert(err, IsNil)
-	c.Assert(repo.types, testutil.Contains, FileType)
-	c.Assert(repo.types, HasLen, 1) // Update this whenever new built-in type is added
-	err = LoadBuiltInTypes(repo)
-	c.Assert(err, ErrorMatches, `cannot add type "file": name already exists`)
-}
 
 func (s *CapabilitySuite) TestString(c *C) {
 	cap := &Capability{Name: "name", Label: "label", Type: FileType}
